pyproject.toml
[tool.poetry]
name = "pilgram"
version = "1.2.1"
description = "library for instagram filters"
authors = ["Akiomi Kamakura <akiomik@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
flake8 = "^5.0.4"
Pillow = "^9.3.0"
pytest-benchmark = "^4.0.0"
pytest-mock = "^3.10.0"
numpy = "^1.23.4"
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
